Blog/requirements.txt
asgiref==3.5.2
cffi==1.15.1
Django==4.1.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
pyasn1==0.4.8
pycparser==2.21
sqlparse==0.4.2
tzdata==2022.2
